Migrate Security cli to use domain directly
