+++ niveshpy/cli/models/report.py
@@ -9,12 +9,14 @@ from decimal import Decimal
 from typing import TYPE_CHECKING, Any, ClassVar, Literal, Self
 
 from niveshpy.cli.models.account import AccountDisplay
-from niveshpy.cli.models.security import (
-    SecurityDisplay,
-    _format_security_type,
+from niveshpy.cli.models.security import SecurityDisplay
+from niveshpy.cli.utils.formatters import (
+    format_date,
+    format_decimal,
+    format_percentage,
     format_security_category,
+    format_security_type,
 )
-from niveshpy.cli.utils.formatters import format_date, format_decimal, format_percentage
 from niveshpy.cli.utils.models import Column
 
 if TYPE_CHECKING:
@@ -146,7 +148,7 @@ class AllocationDisplay:
         """Get the appropriate columns based on whether security type or category is used."""
         columns: list[Column] = []
         if group_by in ("both", "type"):
-            columns.append(Column("security_type", formatter=_format_security_type))
+            columns.append(Column("security_type", formatter=format_security_type))
         if group_by in ("both", "category"):
             columns.append(
                 Column("security_category", formatter=format_security_category)
